Refactor out a Requester class
The Configuration class was doing way too much previously.
It was both setting up the directory and search term, as well as making
an HTTP request to twitter API.

I've extracted a Requester class that handles the API exchange, then
passes the response straight to the Downloader class.

In the future, the Configuration class should probably be refactored
into an Engine classe that executes all the proper methods in other
classes, while all the CLI options get parsed right inside the CLI
class. Instead of the Configuration class invoking a terminal
questionnaire.