Add Typhoeus for concurrent image download
We still need to use regular HTTP request for assembling all the image
links, because every next request depends on the max_id parameter that
we get from the previous request.

The specs should be refactored, as well as the Requester class.
Currently, it is doing too much. There might be a Parser class hidden in
there that should perform all the filtering and such.

Also, stubbing HTTP request in the requester_spec is a good idea and
should be done.