Refactor the Requester class
I have now extracted a separate class called Parser. The Requester still
needs to be cleaned up a bit and the test coverage needs to be improved.

A lot of specs are testing the implementation, instead they should be
testing the behavior, so that all the future refactorings are less
painful.